removed item lag
non chest menu items would have a slight but noticeable lag when selecting them. this is because main ran a check every tick that /cleared the player's check inventory to check for items. now, the only time it will check for items is if the player is near an area effect cloud with a special tag: "chestie". this should further optimize the datapack